change psql lock logic to take lock ID as a parameter
